move shutdown to pkg
