+++ cmd/shortener/main.go
@@ -2,10 +2,8 @@
 package main
 
 import (
-	"context"
 	"database/sql"
 	_ "net/http/pprof"
-	"time"
 
 	_ "github.com/shulganew/shear.git/docs" // docs is generated by Swag CLI, you have to import it.
 	"github.com/shulganew/shear.git/internal/app"
@@ -13,10 +11,9 @@ import (
 	"github.com/shulganew/shear.git/internal/service"
 	"github.com/shulganew/shear.git/internal/web/router"
 	"github.com/shulganew/shear.git/internal/web/server"
+	"github.com/shulganew/shear.git/pkg/shutdown"
 	"go.uber.org/zap"
 )
-
-const timeoutShutdown = time.Second * 20
 
 var (
 	buildVersion string = "N/A"
@@ -38,14 +35,6 @@ func main() {
 
 	// Root app context.
 	ctx, cancel := app.InitContext()
-
-	// Timer hardreset shutdown.
-	context.AfterFunc(ctx, func() {
-		ctx, cancelCtx := context.WithTimeout(context.Background(), timeoutShutdown)
-		defer cancelCtx()
-		<-ctx.Done()
-		zap.S().Fatalln("failed to gracefully shutdown the service")
-	})
 
 	conf := config.NewConfig()
 
@@ -70,10 +59,7 @@ func main() {
 		}()
 	}
 
-	// Error channel.
-	componentsErrs := make(chan error, 1)
-
-	// Use fanIn pattern for storing data from delete requests.
+	// Use fan-In pattern for storing data from delete requests.
 	delCh := make(chan service.DelBatch, 100)
 	defer close(delCh)
 
@@ -83,21 +69,17 @@ func main() {
 	// Create router.
 	rt := router.RouteShear(conf, short, db, del)
 
+	// Error channel.
+	componentsErrs := make(chan error, 1)
+
 	// Start web server.
 	webDone := server.ShortenerServer(ctx, conf, componentsErrs, rt)
 
-	// Update deleted shorts from commont fan-In channel.
+	// Update deleted shorts from common fan-In channel.
 	delDone := app.DeleteShort(ctx, short, delCh)
 
-	// Graceful shutdown.
-	select {
-	// Exit on root context done.
-	case <-ctx.Done():
-	// Exit on errors.
-	case err := <-componentsErrs:
-		zap.S().Errorln("Get server error: ", err)
-		cancel()
-	}
+	//Graceful shutdown.
+	shutdown.Graceful(ctx, cancel, componentsErrs)
 
 	// Waiting http server shuting down.
 	<-webDone
